+++ cheroot/_compat.py
@@ -46,12 +46,10 @@ else:
         # escapes, but without having to prefix it with u'' for Python 2,
         # but no prefix for Python 3.
         if encoding == 'escape':
-            return six.u(
-                re.sub(
-                    r'\\u([0-9a-zA-Z]{4})',
-                    lambda m: six.unichr(int(m.group(1), 16)),
-                    n.decode('ISO-8859-1'),
-                )
+            return re.sub(
+                r'\\u([0-9a-zA-Z]{4})',
+                lambda m: six.unichr(int(m.group(1), 16)),
+                n.decode('ISO-8859-1'),
             )
         # Assume it's already in the given encoding, which for ISO-8859-1
         # is almost always what was intended.
@@ -71,3 +69,25 @@ def assert_native(n):
     """
     if not isinstance(n, str):
         raise TypeError('n must be a native str (got %s)' % type(n).__name__)
+
+
+if six.PY3:
+    """Python 3 has memoryview builtin."""
+    # Python 2.7 has it backported, but socket.write() does
+    # str(memoryview(b'0' * 100)) -> <memory at 0x7fb6913a5588>
+    # instead of accessing it correctly.
+    memoryview = memoryview
+else:
+    """Link memoryview to buffer under Python 2."""
+    memoryview = buffer  # noqa: F821
+
+
+def extract_bytes(mv):
+    """Retrieve bytes out of memoryview/buffer or bytes."""
+    if isinstance(mv, memoryview):
+        return mv.tobytes() if six.PY3 else bytes(mv)
+
+    if isinstance(mv, bytes):
+        return mv
+
+    return ValueError