+++ cheroot/test/test_server.py
@@ -15,7 +15,7 @@ import time
 import pytest
 
 from .._compat import bton
-from ..server import Gateway, HTTPServer
+from ..server import IS_UID_GID_RESOLVABLE, Gateway, HTTPServer
 from ..testing import (
     ANY_INTERFACE_IPV4,
     ANY_INTERFACE_IPV6,
@@ -39,6 +39,11 @@ def make_http_server(bind_addr):
 non_windows_sock_test = pytest.mark.skipif(
     not hasattr(socket, 'AF_UNIX'),
     reason='UNIX domain sockets are only available under UNIX-based OS',
+)
+
+
+http_over_unix_socket = pytest.mark.skip(
+    reason='Test HTTP client is not able to work through UNIX socket currently'
 )
 
 
@@ -73,6 +78,47 @@ def unix_sock_file():
 
     os.close(tmp_sock_fh)
     os.unlink(tmp_sock_fname)
+
+
+def test_prepare_makes_server_ready():
+    """Check that prepare() makes the server ready, and stop() clears it."""
+    httpserver = HTTPServer(
+        bind_addr=(ANY_INTERFACE_IPV4, EPHEMERAL_PORT), gateway=Gateway
+    )
+
+    assert not httpserver.ready
+    assert not httpserver.requests._threads
+
+    httpserver.prepare()
+
+    assert httpserver.ready
+    assert httpserver.requests._threads
+    for thr in httpserver.requests._threads:
+        assert thr.ready
+
+    httpserver.stop()
+
+    assert not httpserver.requests._threads
+    assert not httpserver.ready
+
+
+def test_stop_interrupts_serve():
+    """Check that stop() interrupts running of serve()."""
+    httpserver = HTTPServer(
+        bind_addr=(ANY_INTERFACE_IPV4, EPHEMERAL_PORT), gateway=Gateway
+    )
+
+    httpserver.prepare()
+    serve_thread = threading.Thread(target=httpserver.serve)
+    serve_thread.start()
+
+    serve_thread.join(0.5)
+    assert serve_thread.is_alive()
+
+    httpserver.stop()
+
+    serve_thread.join(0.5)
+    assert not serve_thread.is_alive()
 
 
 @pytest.mark.parametrize('ip_addr', (ANY_INTERFACE_IPV4, ANY_INTERFACE_IPV6))
@@ -119,24 +165,39 @@ class _TestGateway(Gateway):
         return super(_TestGateway, self).respond()
 
 
-@pytest.mark.skip(
-    reason='Test HTTP client is not able to work through UNIX socket currently'
-)
-@non_windows_sock_test
-def test_peercreds_unix_sock(http_server, unix_sock_file):
-    """Check that peercred lookup and resolution work when enabled."""
+@pytest.fixture
+def peercreds_enabled_server_and_client(http_server, unix_sock_file):
+    """Construct a test server with `peercreds_enabled`."""
     httpserver = http_server.send(unix_sock_file)
     httpserver.gateway = _TestGateway
     httpserver.peercreds_enabled = True
+    return httpserver, get_server_client(httpserver)
 
-    testclient = get_server_client(httpserver)
+
+@http_over_unix_socket
+@non_windows_sock_test
+def test_peercreds_unix_sock(peercreds_enabled_server_and_client):
+    """Check that peercred lookup works when enabled."""
+    httpserver, testclient = peercreds_enabled_server_and_client
 
     expected_peercreds = os.getpid(), os.getuid(), os.getgid()
     expected_peercreds = '|'.join(map(str, expected_peercreds))
     assert testclient.get(PEERCRED_IDS_URI) == expected_peercreds
     assert 'RuntimeError' in testclient.get(PEERCRED_TEXTS_URI)
 
+
+@pytest.mark.skipif(
+    not IS_UID_GID_RESOLVABLE,
+    reason='Modules `grp` and `pwd` are not available '
+    'under the current platform',
+)
+@http_over_unix_socket
+@non_windows_sock_test
+def test_peercreds_unix_sock_with_lookup(peercreds_enabled_server_and_client):
+    """Check that peercred resolution works when enabled."""
+    httpserver, testclient = peercreds_enabled_server_and_client
     httpserver.peercreds_resolve_enabled = True
+
     import grp
 
     expected_textcreds = os.getlogin(), grp.getgrgid(os.getgid()).gr_name