+++ cheroot/test/webtest.py
@@ -111,6 +111,7 @@ class WebCase(unittest.TestCase):
 
     scheme = 'http'
     url = None
+    ssl_context = None
 
     status = None
     headers = None
@@ -223,6 +224,7 @@ class WebCase(unittest.TestCase):
             self.HTTP_CONN,
             protocol or self.PROTOCOL,
             raise_subcls,
+            ssl_context=self.ssl_context,
         )
         self.time = time.time() - start
         self.status, self.headers, self.body = result
@@ -495,6 +497,7 @@ def openURL(
     http_conn=http_client.HTTPConnection,
     protocol='HTTP/1.1',
     raise_subcls=None,
+    ssl_context=None,
 ):
     """
     Open the given HTTP resource and return status, headers, and body.
@@ -514,7 +517,10 @@ def openURL(
             if hasattr(http_conn, 'host'):
                 conn = http_conn
             else:
-                conn = http_conn(interface(host), port)
+                kw = {}
+                if ssl_context:
+                    kw['context'] = ssl_context
+                conn = http_conn(interface(host), port, **kw)
 
             conn._http_vsn_str = protocol
             conn._http_vsn = int(''.join([x for x in protocol if x.isdigit()]))