+++ cheroot/workers/threadpool.py
@@ -68,25 +68,25 @@ class WorkerThread(threading.Thread):
         self.stats = {
             'Requests': lambda s: self.requests_seen
             + (
-                (self.start_time is None)
+                self.start_time is None
                 and trueyzero
                 or self.conn.requests_seen
             ),
             'Bytes Read': lambda s: self.bytes_read
             + (
-                (self.start_time is None)
+                self.start_time is None
                 and trueyzero
                 or self.conn.rfile.bytes_read
             ),
             'Bytes Written': lambda s: self.bytes_written
             + (
-                (self.start_time is None)
+                self.start_time is None
                 and trueyzero
                 or self.conn.wfile.bytes_written
             ),
             'Work Time': lambda s: self.work_time
             + (
-                (self.start_time is None)
+                self.start_time is None
                 and trueyzero
                 or time.time() - self.start_time
             ),